Add distributed model executor abstraction
